+++ src/api/smarthome/src/main/java/de/janbellenberg/smarthome/base/helper/security/SecurityLevel.java
@@ -1,0 +1,5 @@
+package de.janbellenberg.smarthome.base.helper.security;
+
+public enum SecurityLevel {
+  ADMIN, USER, BOTH
+}
