--- src/api/smarthome/src/main/java/de/janbellenberg/smarthome/core/helper/security/SecurityLevel.java
@@ -1,5 +1,0 @@
-package de.janbellenberg.smarthome.core.helper.security;
-
-public enum SecurityLevel {
-  ADMIN, USER, BOTH
-}
